Move firestore implementation to the firestore package
